mybatis-plus改造了一下，跑通了，可以瞑目了

+++ src/main/java/boot/spring/po/User.java
@@ -1,7 +1,7 @@
 package boot.spring.po;
 
 public class User {
-	Long uid;
+	int uid;
 	
 	String name;
 	
@@ -9,17 +9,17 @@ public class User {
 		super();
 	}
 
-	public User(Long uid, String name) {
+	public User(int uid, String name) {
 		super();
 		this.uid = uid;
 		this.name = name;
 	}
 
-	public Long getUid() {
+	public int getUid() {
 		return uid;
 	}
 
-	public void setUid(Long uid) {
+	public void setUid(int uid) {
 		this.uid = uid;
 	}
